rbe: use tabs and inline small and simple member methods
