+++ worker/include/RTC/RemoteBitrateEstimator/OveruseDetector.hpp
@@ -17,43 +17,60 @@
 
 namespace RTC {
 
-bool AdaptiveThresholdExperimentIsDisabled();
+	bool AdaptiveThresholdExperimentIsDisabled();
 
-class OveruseDetector {
- public:
-  OveruseDetector();
-  virtual ~OveruseDetector();
+	class OveruseDetector {
+	public:
+		OveruseDetector();
+		virtual ~OveruseDetector();
 
-  // Update the detection state based on the estimated inter-arrival time delta
-  // offset. |timestampDelta| is the delta between the last timestamp which the
-  // estimated offset is based on and the last timestamp on which the last
-  // offset was based on, representing the time between detector updates.
-  // |numOfDeltas| is the number of deltas the offset estimate is based on.
-  // Returns the state after the detection update.
-  BandwidthUsage Detect(double offset,
-                        double timestampDelta,
-                        int numOfDeltas,
-                        int64_t nowMs);
+		// Update the detection state based on the estimated inter-arrival time delta
+		// offset. |timestampDelta| is the delta between the last timestamp which the
+		// estimated offset is based on and the last timestamp on which the last
+		// offset was based on, representing the time between detector updates.
+		// |numOfDeltas| is the number of deltas the offset estimate is based on.
+		// Returns the state after the detection update.
+		BandwidthUsage Detect(double offset,
+				double timestampDelta,
+				int numOfDeltas,
+				int64_t nowMs);
 
-  // Returns the current detector state.
-  BandwidthUsage State() const;
+		// Returns the current detector state.
+		BandwidthUsage State() const;
 
- private:
-  void UpdateThreshold(double modifiedOffset, int64_t nowMs);
-  void InitializeExperiment();
+	private:
+		void UpdateThreshold(double modifiedOffset, int64_t nowMs);
+		void InitializeExperiment();
 
-  bool inExperiment;
-  double kUp;
-  double kDown;
-  double overusingTimeThreshold;
-  double threshold;
-  int64_t lastUpdateMs;
-  double prevOffset;
-  double timeOverUsing;
-  int overuseCounter;
-  BandwidthUsage hypothesis;
+		bool inExperiment;
+		double kUp = 0.0087;
+		double kDown = 0.039;
+		double overusingTimeThreshold = 100;
+		double threshold = 12.5;
+		int64_t lastUpdateMs = -1;
+		double prevOffset = 0.0;
+		double timeOverUsing = -1;
+		int overuseCounter = 0;
+		BandwidthUsage hypothesis = kBwNormal;
 
-};
+	};
+
+	inline
+		OveruseDetector::OveruseDetector() :
+			inExperiment(!AdaptiveThresholdExperimentIsDisabled())
+	{
+		if (this->inExperiment)
+			InitializeExperiment();
+	}
+
+	inline
+		OveruseDetector::~OveruseDetector() {}
+
+	inline
+		BandwidthUsage OveruseDetector::State() const {
+			return this->hypothesis;
+		}
+
 }  // namespace RTC
 
 #endif  // MS_RTC_REMOTE_BITRATE_ESTIMATOR_OVERUSE_DETECTOR_HPP
