+++ worker/include/RTC/RemoteBitrateEstimator/RemoteBitrateEstimator.hpp
@@ -22,73 +22,73 @@
 
 namespace RTC {
 
-// (jmillan) borrowed from webrtc/modules/include/module_common_types.h
-//
-// Interface used by the CallStats class to distribute call statistics.
-// Callbacks will be triggered as soon as the class has been registered to a
-// CallStats object using RegisterStatsObserver.
-class CallStatsObserver {
- public:
-  virtual void OnRttUpdate(int64_t avg_rtt_ms, int64_t max_rtt_ms) = 0;
+	// (jmillan) borrowed from webrtc/modules/include/module_common_types.h
+	//
+	// Interface used by the CallStats class to distribute call statistics.
+	// Callbacks will be triggered as soon as the class has been registered to a
+	// CallStats object using RegisterStatsObserver.
+	class CallStatsObserver {
+		public:
+			virtual void OnRttUpdate(int64_t avg_rtt_ms, int64_t max_rtt_ms) = 0;
 
-  virtual ~CallStatsObserver() {}
-};
+			virtual ~CallStatsObserver() {}
+	};
 
-class RemoteBitrateEstimator : public CallStatsObserver {
- public:
-   // Used to signal changes in bitrate estimates for the incoming streams.
-   class Listener
-   {
-     public:
-       // Called when a receive channel group has a new bitrate estimate for the incoming streams.
-       virtual void onReceiveBitrateChanged(const std::vector<uint32_t>& ssrcs, uint32_t bitrate) = 0;
-   };
+	class RemoteBitrateEstimator : public CallStatsObserver {
+		public:
+			// Used to signal changes in bitrate estimates for the incoming streams.
+			class Listener
+			{
+				public:
+					// Called when a receive channel group has a new bitrate estimate for the incoming streams.
+					virtual void onReceiveBitrateChanged(const std::vector<uint32_t>& ssrcs, uint32_t bitrate) = 0;
+			};
 
- public:
-  virtual ~RemoteBitrateEstimator() {}
+		public:
+			virtual ~RemoteBitrateEstimator() {}
 
-  // Called for each incoming packet. Updates the incoming payload bitrate
-  // estimate and the over-use detector. If an over-use is detected the
-  // remote bitrate estimate will be updated. Note that |payload_size| is the
-  // packet size excluding headers.
-  // Note that |arrival_time_ms| can be of an arbitrary time base.
-  virtual void IncomingPacket(int64_t arrival_time_ms,
-                              size_t payload_size,
-                              const RtpPacket& packet,
-                              const uint8_t* extensionHeaderValue) = 0;
+			// Called for each incoming packet. Updates the incoming payload bitrate
+			// estimate and the over-use detector. If an over-use is detected the
+			// remote bitrate estimate will be updated. Note that |payload_size| is the
+			// packet size excluding headers.
+			// Note that |arrival_time_ms| can be of an arbitrary time base.
+			virtual void IncomingPacket(int64_t arrival_time_ms,
+					size_t payload_size,
+					const RtpPacket& packet,
+					const uint8_t* extensionHeaderValue) = 0;
 
-  // Removes all data for |ssrc|.
-  virtual void RemoveStream(uint32_t ssrc) = 0;
+			// Removes all data for |ssrc|.
+			virtual void RemoveStream(uint32_t ssrc) = 0;
 
-  // Returns true if a valid estimate exists and sets |bitrate_bps| to the
-  // estimated payload bitrate in bits per second. |ssrcs| is the list of ssrcs
-  // currently being received and of which the bitrate estimate is based upon.
-  virtual bool LatestEstimate(std::vector<uint32_t>* ssrcs,
-                              uint32_t* bitrate_bps) const = 0;
+			// Returns true if a valid estimate exists and sets |bitrate_bps| to the
+			// estimated payload bitrate in bits per second. |ssrcs| is the list of ssrcs
+			// currently being received and of which the bitrate estimate is based upon.
+			virtual bool LatestEstimate(std::vector<uint32_t>* ssrcs,
+					uint32_t* bitrate_bps) const = 0;
 
-  virtual void SetMinBitrate(int min_bitrate_bps) = 0;
+			virtual void SetMinBitrate(int min_bitrate_bps) = 0;
 
-  // (jmillan) borrowed from webrtc/modules/include/module.h.
-  //
-  // Returns the number of milliseconds until the module wants a worker
-  // thread to call Process.
-  // This method is called on the same worker thread as Process will
-  // be called on.
-  // TODO(tommi): Almost all implementations of this function, need to know
-  // the current tick count.  Consider passing it as an argument.  It could
-  // also improve the accuracy of when the next callback occurs since the
-  // thread that calls Process() will also have it's tick count reference
-  // which might not match with what the implementations use.
-  virtual int64_t TimeUntilNextProcess() = 0;
+			// (jmillan) borrowed from webrtc/modules/include/module.h.
+			//
+			// Returns the number of milliseconds until the module wants a worker
+			// thread to call Process.
+			// This method is called on the same worker thread as Process will
+			// be called on.
+			// TODO(tommi): Almost all implementations of this function, need to know
+			// the current tick count.  Consider passing it as an argument.  It could
+			// also improve the accuracy of when the next callback occurs since the
+			// thread that calls Process() will also have it's tick count reference
+			// which might not match with what the implementations use.
+			virtual int64_t TimeUntilNextProcess() = 0;
 
-  // Process any pending tasks such as timeouts.
-  // Called on a worker thread.
-  virtual void Process() = 0;
+			// Process any pending tasks such as timeouts.
+			// Called on a worker thread.
+			virtual void Process() = 0;
 
- protected:
-  static const int64_t kProcessIntervalMs = 500;
-  static const int64_t kStreamTimeOutMs = 2000;
-};
+		protected:
+			static const int64_t kProcessIntervalMs = 500;
+			static const int64_t kStreamTimeOutMs = 2000;
+	};
 
 }  // namespace RTC
